TIP-1159: Rename product to completenes mask
