+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/completeness.yml
@@ -13,7 +13,7 @@ services:
     pim_catalog.completeness.calculator:
         class: 'Akeneo\Pim\Enrichment\Component\Product\Completeness\CompletenessCalculator'
         arguments:
-            - '@akeneo.pim.enrichment.completeness.query.get_products'
+            - '@akeneo.pim.enrichment.completeness.query.get_product_masks'
             - '@akeneo.pim.enrichment.completeness.query.get_family_masks'
 
     pim_catalog.completeness.generator:
