+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/completeness_queries.yml
@@ -1,6 +1,6 @@
 services:
-    akeneo.pim.enrichment.completeness.query.get_products:
-        class: 'Akeneo\Pim\Enrichment\Bundle\Product\Query\Sql\Completeness\SqlGetProducts'
+    akeneo.pim.enrichment.completeness.query.get_product_masks:
+        class: Akeneo\Pim\Enrichment\Bundle\Product\Query\Sql\Completeness\SqlGetCompletenessProductMasks
         arguments:
             - '@database_connection'
